requirements.txt
tqdm
Pillow
numpy
scikit-learn
matplotlib
diffusers